remove opencv2/cvconfig.h dependency.
cvconfig.h is not installed in the regular opencv folder location on ubuntu, hence it is not found by the cmake macro "findOpenCV" and the include links are not properly set.

I thought about rewriting the cmake macro to inlucde the file direcotry and link it correctly, but then I found out that it is not needed to build any of the files, therefore
I removed the dependency.